Frontends/VirtualBox: get rid of vcc warnings
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@35448 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIFrameBuffer.h
@@ -5,7 +5,7 @@
  */
 
 /*
- * Copyright (C) 2010 Oracle Corporation
+ * Copyright (C) 2010-2011 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -235,8 +235,8 @@ protected:
 
     UIMachineView *m_pMachineView;
     RTCRITSECT m_critSect;
-    int m_width;
-    int m_height;
+    ulong m_width;
+    ulong m_height;
     QSize m_scaledSize;
     int64_t m_WinId;
     bool m_fIsDeleted;
